requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
discord-ext-voice-recv==0.4.2a145
discord.py==2.4.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
gTTS==2.5.4
idna==3.10
Jinja2==3.1.5
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
propcache==0.2.1
pycparser==2.22
pydub==0.25.1
PyNaCl==1.5.0
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
srt==3.5.3
sympy==1.13.1
tiktoken==0.8.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
vosk==0.3.45
websockets==14.2
yarl==1.18.3
